refactor(web): move src/domains/assistant → src/assistant (LUM-1753.3)
The assistant isn't a feature alongside chat / settings / library —
it's the central concept the whole app is built around. Every
feature depends on it; it depends on no feature. The directory
layout should match that relationship.

Promotes `src/domains/assistant/` to a top-level `src/assistant/`
folder, sibling of the technical-purpose dirs (`hooks/`, `stores/`,
`lib/`, `utils/`, `components/`, `runtime/`) and the feature dir
(`domains/`).

What changed:
- `git mv src/domains/assistant src/assistant` (history preserved).
- Codemod: `@/domains/assistant/...` → `@/assistant/...` across
  the codebase. 28 imports updated.
- Cross-domain allow-list regenerated: 148 → 120 imports (-28),
  109 → 93 files (-16). All formerly inbound assistant imports
  drop out — they're no longer cross-domain because the target
  isn't a domain anymore.
- CONVENTIONS.md gets a new top-level row documenting the
  category: business-domain code that every feature may depend
  on but that depends on no feature. New top-level
  business-concept folders require explicit team approval (we
  don't want this category drifting open).
- README directory diagram updated.

Decision rationale recorded in LUM-1755.
