+++ apps/web/src/assistant/llm-model-catalog.ts
