+++ apps/web/src/domains/chat/api/event-types.ts
@@ -6,7 +6,7 @@
  * consumed by event-parser.ts and the stream handler domain modules.
  */
 
-import type { DiskPressureStatus } from "@/domains/assistant/types.js";
+import type { DiskPressureStatus } from "@/assistant/types.js";
 import type { Surface } from "@/domains/chat/types/types.js";
 import type { SyncChangedEvent } from "@/lib/sync/types.js";
