+++ apps/web/src/domains/chat/hooks/use-stream-event-handler.ts
@@ -12,7 +12,7 @@ import { useConversationStore } from "@/domains/conversations/conversation-store
 import type { ContextWindowUsage } from "@/domains/chat/components/context-window-indicator.js";
 import type { DisplayMessage } from "@/domains/chat/utils/reconcile.js";
 import { useTurnStore } from "@/domains/messaging/turn-store.js";
-import type { DiskPressureStatusEventPayload } from "@/domains/assistant/use-disk-pressure-monitor.js";
+import type { DiskPressureStatusEventPayload } from "@/assistant/use-disk-pressure-monitor.js";
 import {
   recordChatDiagnostic,
   summarizeAssistantEvent,
