+++ apps/web/src/domains/chat/utils/stream-handlers/types.ts
@@ -7,7 +7,7 @@ import type { QueryClient } from "@tanstack/react-query";
 import type { ContextWindowUsage } from "@/domains/chat/components/context-window-indicator.js";
 import type { DisplayMessage } from "@/domains/chat/utils/reconcile.js";
 import type { TurnActions, TurnState } from "@/domains/messaging/turn-store.js";
-import type { DiskPressureStatusEventPayload } from "@/domains/assistant/use-disk-pressure-monitor.js";
+import type { DiskPressureStatusEventPayload } from "@/assistant/use-disk-pressure-monitor.js";
 import type { ChatError, PendingQuestionState } from "@/domains/chat/types.js";
 import type { ChatEventStream } from "@/domains/chat/api/stream.js";
