+++ apps/web/src/domains/settings/ai/ai-page.tsx
@@ -46,14 +46,14 @@ import { useFeatureFlagStore } from "@/lib/feature-flags/feature-flag-store.js";
 import {
   type LlmCatalogModel,
   PROVIDER_DISPLAY_NAMES,
-} from "@/domains/assistant/llm-model-catalog.js";
+} from "@/assistant/llm-model-catalog.js";
 import {
   WEB_SEARCH_BYOK_PROVIDER_IDS,
   WEB_SEARCH_PROVIDER_DISPLAY_NAMES,
   WEB_SEARCH_PROVIDER_IDS,
   WEB_SEARCH_PROVIDER_KEY_PLACEHOLDERS,
   WEB_SEARCH_PROVIDER_KEY_STORAGE,
-} from "@/domains/assistant/generated/web-search-provider-catalog.gen.js";
+} from "@/assistant/generated/web-search-provider-catalog.gen.js";
 import { routes } from "@/utils/routes.js";
 import { assistantDaemonConfigQueryKey } from "@/lib/sync/query-tags.js";
 import { synthesizeTTS } from "@/domains/voice/tts-synthesize.js";
@@ -100,7 +100,7 @@ export const TOKEN_SLIDER_STEP_TOKENS = 1_000;
 export const DEFAULT_CONTEXT_WINDOW_BUDGET_TOKENS = 200_000;
 
 // Web-search provider list, display names, and key placeholders live in
-// `domains/assistant/generated/web-search-provider-catalog.gen.ts`.
+// `assistant/generated/web-search-provider-catalog.gen.ts`.
 
 // ---------------------------------------------------------------------------
 // TTS / STT provider catalogs — mirror the macOS client catalogs and the
