+++ apps/web/src/domains/settings/pages/integrations-page.tsx
@@ -18,7 +18,7 @@ import { Button } from "@vellum/design-library/components/button";
 import { Input } from "@vellum/design-library/components/input";
 import { assistantsOauthConnectionsListOptions } from "@/generated/api/@tanstack/react-query.gen.js";
 import type { OAuthConnection } from "@/generated/api/types.gen.js";
-import { type Assistant, getAssistant } from "@/domains/assistant/api.js";
+import { type Assistant, getAssistant } from "@/assistant/api.js";
 import {
   fetchOAuthProviders,
   type OAuthProviderSummary,
